requirements.txt
anthropic
boto3
docker
gitpython
